pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "depth-app-ml"
version = "1.0.0"
description = "Depth estimation model exporter"
requires-python = "==3.10.*"
dependencies = [
    "torch==2.2.0",
    "torchvision==0.17.0",
    "transformers==4.36.2",
    "huggingface-hub==0.20.3",
    "numpy==1.24.3",
    "onnx==1.15.0",
    "onnxruntime==1.17.1",
    "Pillow==10.1.0",
    "opencv-python==4.8.1.78",
    "tflite-runtime==2.14.0; platform_system == 'Linux'",
]

[project.optional-dependencies]
cuda = [
    "torch==2.2.0+cu118",
    "torchvision==0.17.0+cu118",
]
